fixing the styles!
