+++ public/translations/de.csv
@@ -11,6 +11,7 @@ offense.coverage.edit,Edit list,
 offense.coverage.weakness,Weak,
 offense.coverage.normal,Normal,
 offense.coverage.resistance,Resist,
+offense.teraPokemon,Tera Pokémon,
 offense.coverageList.empty,No Pokémon found or no types selected,
 offense.coverageList.weakness.heading,Weakness Coverage List,
 offense.coverageList.weakness.description,These Pokémon are weak to at least one of the selected types.,
@@ -44,6 +45,7 @@ defense.chooseSecond,Choose second type,Zweiten Typ wählen
 defense.chooseThird,Choose third type,Dritten Typ wählen
 defense.takesXFrom,Takes {{ x }} from,Nimmt {{ x }} Schaden von
 defense.chooseAbility,Choose ability,
+defense.chooseTeraType,Choose tera type,
 defense.abilityNames.none,None,
 defense.abilityNames.wonder_guard,Wonder Guard,Wonder Guard
 defense.abilityNames.delta_stream,Delta Stream,Delta Stream
@@ -186,7 +188,7 @@ types.ice,Ice,Eis
 types.dragon,Dragon,Drache
 types.dark,Dark,Unlicht
 types.fairy,Fairy,Fee
-types.stellar,Stellar,
+types.stellar,Stellar,Stellar
 types.none,None,Kein
 banners.updateReady.description,An update is available,Ein Update ist verfügbar
 banners.updateReady.whatsNew,What’s new,
