+++ public/translations/es.csv
@@ -11,6 +11,7 @@ offense.coverage.edit,Edit list,Modificar lista
 offense.coverage.weakness,Weak,Debilidad
 offense.coverage.normal,Normal,Normal
 offense.coverage.resistance,Resist,Resistencia
+offense.teraPokemon,Tera Pokémon,
 offense.coverageList.empty,No Pokémon found or no types selected,No se han encontrado Pokémon o no se han seleccionado tipos
 offense.coverageList.weakness.heading,Weakness Coverage List,Lista de debilidad
 offense.coverageList.weakness.description,These Pokémon are weak to at least one of the selected types.,Estos Pokémon son débiles a al menos uno de los tipos seleccionados.
@@ -44,6 +45,7 @@ defense.chooseSecond,Choose second type,Elige el tipo secundario
 defense.chooseThird,Choose third type,Elige el tipo terciario
 defense.takesXFrom,Takes {{ x }} from,Sufre {{ x }} por
 defense.chooseAbility,Choose ability,Elige Habilidad
+defense.chooseTeraType,Choose tera type,
 defense.abilityNames.none,None,Ninguna
 defense.abilityNames.wonder_guard,Wonder Guard,Superguarda
 defense.abilityNames.delta_stream,Delta Stream,Ráfaga Delta
@@ -186,7 +188,7 @@ types.ice,Ice,Hielo
 types.dragon,Dragon,Dragón
 types.dark,Dark,Siniestro
 types.fairy,Fairy,Hada
-types.stellar,Stellar,
+types.stellar,Stellar,Astral
 types.none,None,Ninguno
 banners.updateReady.description,An update is available,Hay una actualización disponible
 banners.updateReady.whatsNew,What’s new,Novedades
