+++ public/translations/zh-Hant.csv
@@ -11,6 +11,7 @@ offense.coverage.edit,Edit list,編輯列表
 offense.coverage.weakness,Weak,弱點
 offense.coverage.normal,Normal,正常
 offense.coverage.resistance,Resist,抵抗
+offense.teraPokemon,Tera Pokémon,
 offense.coverageList.empty,No Pokémon found or no types selected,未找到寶可夢或未選擇屬性
 offense.coverageList.weakness.heading,Weakness Coverage List,弱點涵蓋列表
 offense.coverageList.weakness.description,These Pokémon are weak to at least one of the selected types.,這些寶可夢對所選擇的屬性至少有一個弱點。
@@ -44,6 +45,7 @@ defense.chooseSecond,Choose second type,選擇第二個屬性
 defense.chooseThird,Choose third type,選擇第三個屬性
 defense.takesXFrom,Takes {{ x }} from,從中取 {{ x }}
 defense.chooseAbility,Choose ability,選擇技能
+defense.chooseTeraType,Choose tera type,
 defense.abilityNames.none,None,無
 defense.abilityNames.wonder_guard,Wonder Guard,奇跡防守
 defense.abilityNames.delta_stream,Delta Stream,德爾塔氣流
@@ -186,7 +188,7 @@ types.ice,Ice,冰
 types.dragon,Dragon,龍
 types.dark,Dark,黑暗
 types.fairy,Fairy,妖精
-types.stellar,Stellar,
+types.stellar,Stellar,星晶
 types.none,None,無
 banners.updateReady.description,An update is available,有更新可用
 banners.updateReady.whatsNew,What’s new,有什麼新功能
