+++ src/Badge.module.css
@@ -1,0 +1,19 @@
+.badgeContainer {
+  padding: 3px;
+  box-shadow: inset 0 0 0 1px var(--color-vibrant-border);
+}
+
+.badgeContainer[data-type="stellar"] {
+  background: var(--stellar-background-linear);
+}
+
+.badgeLabel {
+  padding-left: 1rem;
+  padding-right: 1rem;
+  background: var(--type-color-bg);
+}
+
+.badgeLabel[data-type="stellar"] {
+  background: hsl(0 0% 25% / 50%);
+  backdrop-filter: blur(1px);
+}
