+++ src/Badge.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import { useTranslation } from "react-i18next";
 import { typeColor, typeColorBG, typeColorBorder } from "./colors";
 import { Type } from "./data-types";
+import styles from "./Badge.module.css";
 
 interface BadgeProps {
   type: Type;
@@ -14,27 +15,23 @@ export function Badge({ type }: BadgeProps) {
     <div
       className={classNames(
         "type-bg",
-        "ba border-vibrant",
         "br2",
         "b f5 lh-title tc",
-        type === Type.stellar && "type-stellar"
+        styles.badgeContainer
       )}
+      data-type={type}
       style={{
-        padding: 2,
         ["--type-color" as any]: typeColor(type),
+        ["--type-color-bg" as any]: typeColorBG(type),
+        ["--type-color-border" as any]: typeColorBorder(type),
       }}
     >
       <div
         className={classNames(
           "br1 ba b--transparent white truncate",
-          type === Type.stellar ? "type-stellar-label" : "type-any-label"
+          styles.badgeLabel
         )}
-        style={{
-          paddingLeft: 4,
-          paddingRight: 4,
-          ["--type-color-bg" as any]: typeColorBG(type),
-          ["--type-color-border" as any]: typeColorBorder(type),
-        }}
+        data-type={type}
       >
         {t(`types.${type}`)}
       </div>
