+++ src/style.css
@@ -26,6 +26,61 @@
   --type-text-shadow: 0 1px 0 black, 0 0 1px rgb(0 0 0 / 60%),
     0 0 2px rgb(0 0 0 / 70%), 0 0 3px rgb(0 0 0 / 80%), 0 0 4px rgb(0 0 0 / 90%);
 
+  /* Created bey stellar-styles.ts */
+  --stellar-background-linear: linear-gradient(
+    -225deg,
+    hsl(0 100% 50%) 0 calc(6.25% + 1px),
+    hsl(22.5 100% 50%) calc(6.25% - 1px) calc(12.5% + 1px),
+    hsl(45 100% 50%) calc(12.5% - 1px) calc(18.75% + 1px),
+    hsl(67.5 100% 50%) calc(18.75% - 1px) calc(25% + 1px),
+    hsl(90 100% 50%) calc(25% - 1px) calc(31.25% + 1px),
+    hsl(112.5 100% 50%) calc(31.25% - 1px) calc(37.5% + 1px),
+    hsl(135 100% 50%) calc(37.5% - 1px) calc(43.75% + 1px),
+    hsl(157.5 100% 50%) calc(43.75% - 1px) calc(50% + 1px),
+    hsl(180 100% 50%) calc(50% - 1px) calc(56.25% + 1px),
+    hsl(202.5 100% 50%) calc(56.25% - 1px) calc(62.5% + 1px),
+    hsl(225 100% 50%) calc(62.5% - 1px) calc(68.75% + 1px),
+    hsl(247.5 100% 50%) calc(68.75% - 1px) calc(75% + 1px),
+    hsl(270 100% 50%) calc(75% - 1px) calc(81.25% + 1px),
+    hsl(292.5 100% 50%) calc(81.25% - 1px) calc(87.5% + 1px),
+    hsl(315 100% 50%) calc(87.5% - 1px) calc(93.75% + 1px)
+  );
+  --stellar-background-conic: conic-gradient(
+    hsl(0 100% 50%) 0deg 18.5deg,
+    hsl(22.5 100% 50%) 22.5deg 41deg,
+    hsl(45 100% 50%) 45deg 63.5deg,
+    hsl(67.5 100% 50%) 67.5deg 86deg,
+    hsl(90 100% 50%) 90deg 108.5deg,
+    hsl(112.5 100% 50%) 112.5deg 131deg,
+    hsl(135 100% 50%) 135deg 153.5deg,
+    hsl(157.5 100% 50%) 157.5deg 176deg,
+    hsl(180 100% 50%) 180deg 198.5deg,
+    hsl(202.5 100% 50%) 202.5deg 221deg,
+    hsl(225 100% 50%) 225deg 243.5deg,
+    hsl(247.5 100% 50%) 247.5deg 266deg,
+    hsl(270 100% 50%) 270deg 288.5deg,
+    hsl(292.5 100% 50%) 292.5deg 311deg,
+    hsl(315 100% 50%) 315deg 333.5deg
+  );
+  --stellar-background-dark: linear-gradient(
+    -225deg,
+    hsl(0 100% 20%) 0 calc(6.25% + 1px),
+    hsl(22.5 100% 20%) calc(6.25% - 1px) calc(12.5% + 1px),
+    hsl(45 100% 20%) calc(12.5% - 1px) calc(18.75% + 1px),
+    hsl(67.5 100% 20%) calc(18.75% - 1px) calc(25% + 1px),
+    hsl(90 100% 20%) calc(25% - 1px) calc(31.25% + 1px),
+    hsl(112.5 100% 20%) calc(31.25% - 1px) calc(37.5% + 1px),
+    hsl(135 100% 20%) calc(37.5% - 1px) calc(43.75% + 1px),
+    hsl(157.5 100% 20%) calc(43.75% - 1px) calc(50% + 1px),
+    hsl(180 100% 20%) calc(50% - 1px) calc(56.25% + 1px),
+    hsl(202.5 100% 20%) calc(56.25% - 1px) calc(62.5% + 1px),
+    hsl(225 100% 20%) calc(62.5% - 1px) calc(68.75% + 1px),
+    hsl(247.5 100% 20%) calc(68.75% - 1px) calc(75% + 1px),
+    hsl(270 100% 20%) calc(75% - 1px) calc(81.25% + 1px),
+    hsl(292.5 100% 20%) calc(81.25% - 1px) calc(87.5% + 1px),
+    hsl(315 100% 20%) calc(87.5% - 1px) calc(93.75% + 1px)
+  );
+
   scrollbar-color: var(--color-scroll-fg) var(--color-scroll-bg);
   overflow-y: scroll;
   color-scheme: dark light;
@@ -286,7 +341,7 @@ button {
 
 .type-any-label {
   background: var(--type-color-bg);
-  border-color: var(--type-color-border);
+  box-shadow: 0 0 0 1px var(--type-color-border);
 }
 
 .select-none {
